Make the YR_ARENA_REF struct to be 64-bits long without any padding.
When we have assignments like ref = YR_ARENA_NULL_REF we don't want the resulting struct "ref" to have padding bytes that can be filled with random bytes from memory. If the YR_ARENA_REF has some padding we can't make sure that result from such assignments don't have random values in the padding bytes.

+++ libyara/include/yara/arena.h
@@ -40,7 +40,7 @@ SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #define YR_ARENA_FILE_VERSION  17
 
 #define YR_ARENA_NULL_REF  \
-    (YR_ARENA_REF){ UINT8_MAX, UINT32_MAX }
+    (YR_ARENA_REF){ UINT32_MAX, UINT32_MAX }
 
 typedef uint32_t yr_arena_off_t;
 
@@ -63,7 +63,7 @@ typedef struct YR_RELOC YR_RELOC;
 
 struct YR_ARENA_REF
 {
-  uint8_t buffer_id;
+  uint32_t buffer_id;
   uint32_t offset;
 };
